+++ crates/rust-client/src/notes/import.rs
@@ -7,7 +7,7 @@ use miden_objects::{
 use winter_maybe_async::maybe_await;
 
 use crate::{
-    store::{ExpectedNoteState, InputNoteRecord, NoteState},
+    store::{input_note_states::ExpectedNoteState, InputNoteRecord, InputNoteState},
     sync::NoteTagRecord,
     Client, ClientError,
 };
@@ -48,7 +48,8 @@ impl<R: FeltRng> Client<R> {
         };
 
         if let Some(note) = note {
-            if let NoteState::Expected(ExpectedNoteState { tag: Some(tag), .. }) = note.state() {
+            if let InputNoteState::Expected(ExpectedNoteState { tag: Some(tag), .. }) = note.state()
+            {
                 maybe_await!(self
                     .store
                     .add_note_tag(NoteTagRecord::with_note_source(*tag, note.id())))?;
@@ -122,10 +123,10 @@ impl<R: FeltRng> Client<R> {
         }
     }
 
-    /// Builds a note record from the note and inclusion proof. If a note with the same id was
+    /// Builds a note record from the note and inclusion proof. If a note with the same ID was
     /// already stored it is passed via `previous_note` so it can be updated. The note's
     /// nullifier is used to determine if the note has been consumed in the node and gives it
-    /// the correct status.
+    /// the correct state.
     ///
     /// If the note is not consumed and it was committed in the past relative to the client, then
     /// the MMR for the relevant block is fetched from the node and stored.