+++ crates/rust-client/src/store/note_record/input_note_record/states/processing_authenticated.rs
@@ -1,41 +1,49 @@
 use alloc::string::ToString;
 
 use miden_objects::{
+    accounts::AccountId,
     notes::{NoteId, NoteInclusionProof, NoteMetadata},
     transaction::TransactionId,
-    BlockHeader,
+    BlockHeader, Digest,
 };
 
 use super::{
-    ConsumedExternalNoteState, ConsumedUnauthenticatedLocalNoteState, NoteState, NoteStateHandler,
-    NoteSubmissionData,
+    ConsumedAuthenticatedLocalNoteState, ConsumedExternalNoteState, InputNoteState,
+    NoteStateHandler, NoteSubmissionData,
 };
 use crate::store::NoteRecordError;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct ProcessingUnauthenticatedNoteState {
+pub struct ProcessingAuthenticatedNoteState {
     /// Metadata associated with the note, including sender, note type, tag and other additional
     /// information.
     pub metadata: NoteMetadata,
-    /// Block height after which the note is expected to be committed.
-    pub after_block_num: u32,
+    /// Inclusion proof for the note inside the chain block.
+    pub inclusion_proof: NoteInclusionProof,
+    /// Root of the note tree inside the block that verifies the note inclusion proof.
+    pub block_note_root: Digest,
     /// Information about the submission of the note.
     pub submission_data: NoteSubmissionData,
 }
 
-impl NoteStateHandler for ProcessingUnauthenticatedNoteState {
+impl NoteStateHandler for ProcessingAuthenticatedNoteState {
     fn inclusion_proof_received(
         &self,
-        _inclusion_proof: NoteInclusionProof,
-        _metadata: NoteMetadata,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
+        inclusion_proof: NoteInclusionProof,
+        metadata: NoteMetadata,
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
+        if self.inclusion_proof != inclusion_proof || self.metadata != metadata {
+            return Err(NoteRecordError::StateTransitionError(
+                "Inclusion proof or metadata do not match the expected values".to_string(),
+            ));
+        }
         Ok(None)
     }
 
     fn consumed_externally(
         &self,
         nullifier_block_height: u32,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Ok(Some(ConsumedExternalNoteState { nullifier_block_height }.into()))
     }
 
@@ -43,15 +51,15 @@ impl NoteStateHandler for ProcessingUnauthenticatedNoteState {
         &self,
         _note_id: NoteId,
         _block_header: BlockHeader,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Ok(None)
     }
 
     fn consumed_locally(
         &self,
-        _consumer_account: miden_objects::accounts::AccountId,
-        _consumer_transaction: miden_objects::transaction::TransactionId,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
+        _consumer_account: AccountId,
+        _consumer_transaction: TransactionId,
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Err(NoteRecordError::NoteNotConsumable("Note being consumed".to_string()))
     }
 
@@ -59,7 +67,7 @@ impl NoteStateHandler for ProcessingUnauthenticatedNoteState {
         &self,
         transaction_id: TransactionId,
         block_height: u32,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
         if transaction_id != self.submission_data.consumer_transaction {
             return Err(NoteRecordError::StateTransitionError(
                 "Transaction ID does not match the expected value".to_string(),
@@ -67,8 +75,10 @@ impl NoteStateHandler for ProcessingUnauthenticatedNoteState {
         }
 
         Ok(Some(
-            ConsumedUnauthenticatedLocalNoteState {
+            ConsumedAuthenticatedLocalNoteState {
                 metadata: self.metadata,
+                inclusion_proof: self.inclusion_proof.clone(),
+                block_note_root: self.block_note_root,
                 nullifier_block_height: block_height,
                 submission_data: self.submission_data,
             }
@@ -81,7 +91,7 @@ impl NoteStateHandler for ProcessingUnauthenticatedNoteState {
     }
 
     fn inclusion_proof(&self) -> Option<&NoteInclusionProof> {
-        None
+        Some(&self.inclusion_proof)
     }
 
     fn consumer_transaction_id(&self) -> Option<&TransactionId> {
@@ -89,31 +99,34 @@ impl NoteStateHandler for ProcessingUnauthenticatedNoteState {
     }
 }
 
-impl miden_tx::utils::Serializable for ProcessingUnauthenticatedNoteState {
+impl miden_tx::utils::Serializable for ProcessingAuthenticatedNoteState {
     fn write_into<W: miden_tx::utils::ByteWriter>(&self, target: &mut W) {
         self.metadata.write_into(target);
-        self.after_block_num.write_into(target);
+        self.inclusion_proof.write_into(target);
+        self.block_note_root.write_into(target);
         self.submission_data.write_into(target);
     }
 }
 
-impl miden_tx::utils::Deserializable for ProcessingUnauthenticatedNoteState {
+impl miden_tx::utils::Deserializable for ProcessingAuthenticatedNoteState {
     fn read_from<R: miden_tx::utils::ByteReader>(
         source: &mut R,
     ) -> Result<Self, miden_tx::utils::DeserializationError> {
         let metadata = NoteMetadata::read_from(source)?;
-        let after_block_num = u32::read_from(source)?;
+        let inclusion_proof = NoteInclusionProof::read_from(source)?;
+        let block_note_root = Digest::read_from(source)?;
         let submission_data = NoteSubmissionData::read_from(source)?;
-        Ok(ProcessingUnauthenticatedNoteState {
+        Ok(ProcessingAuthenticatedNoteState {
             metadata,
-            after_block_num,
+            inclusion_proof,
+            block_note_root,
             submission_data,
         })
     }
 }
 
-impl From<ProcessingUnauthenticatedNoteState> for NoteState {
-    fn from(state: ProcessingUnauthenticatedNoteState) -> Self {
-        NoteState::ProcessingUnauthenticated(state)
+impl From<ProcessingAuthenticatedNoteState> for InputNoteState {
+    fn from(state: ProcessingAuthenticatedNoteState) -> Self {
+        InputNoteState::ProcessingAuthenticated(state)
     }
 }