+++ crates/rust-client/src/store/note_record/input_note_record/states/processing_unauthenticated.rs
@@ -1,49 +1,41 @@
 use alloc::string::ToString;
 
 use miden_objects::{
-    accounts::AccountId,
     notes::{NoteId, NoteInclusionProof, NoteMetadata},
     transaction::TransactionId,
-    BlockHeader, Digest,
+    BlockHeader,
 };
 
 use super::{
-    ConsumedAuthenticatedLocalNoteState, ConsumedExternalNoteState, NoteState, NoteStateHandler,
-    NoteSubmissionData,
+    ConsumedExternalNoteState, ConsumedUnauthenticatedLocalNoteState, InputNoteState,
+    NoteStateHandler, NoteSubmissionData,
 };
 use crate::store::NoteRecordError;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct ProcessingAuthenticatedNoteState {
+pub struct ProcessingUnauthenticatedNoteState {
     /// Metadata associated with the note, including sender, note type, tag and other additional
     /// information.
     pub metadata: NoteMetadata,
-    /// Inclusion proof for the note inside the chain block.
-    pub inclusion_proof: NoteInclusionProof,
-    /// Root of the note tree inside the block that verifies the note inclusion proof.
-    pub block_note_root: Digest,
+    /// Block height after which the note is expected to be committed.
+    pub after_block_num: u32,
     /// Information about the submission of the note.
     pub submission_data: NoteSubmissionData,
 }
 
-impl NoteStateHandler for ProcessingAuthenticatedNoteState {
+impl NoteStateHandler for ProcessingUnauthenticatedNoteState {
     fn inclusion_proof_received(
         &self,
-        inclusion_proof: NoteInclusionProof,
-        metadata: NoteMetadata,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
-        if self.inclusion_proof != inclusion_proof || self.metadata != metadata {
-            return Err(NoteRecordError::StateTransitionError(
-                "Inclusion proof or metadata do not match the expected values".to_string(),
-            ));
-        }
+        _inclusion_proof: NoteInclusionProof,
+        _metadata: NoteMetadata,
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Ok(None)
     }
 
     fn consumed_externally(
         &self,
         nullifier_block_height: u32,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Ok(Some(ConsumedExternalNoteState { nullifier_block_height }.into()))
     }
 
@@ -51,15 +43,15 @@ impl NoteStateHandler for ProcessingAuthenticatedNoteState {
         &self,
         _note_id: NoteId,
         _block_header: BlockHeader,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Ok(None)
     }
 
     fn consumed_locally(
         &self,
-        _consumer_account: AccountId,
-        _consumer_transaction: TransactionId,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
+        _consumer_account: miden_objects::accounts::AccountId,
+        _consumer_transaction: miden_objects::transaction::TransactionId,
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Err(NoteRecordError::NoteNotConsumable("Note being consumed".to_string()))
     }
 
@@ -67,7 +59,7 @@ impl NoteStateHandler for ProcessingAuthenticatedNoteState {
         &self,
         transaction_id: TransactionId,
         block_height: u32,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
         if transaction_id != self.submission_data.consumer_transaction {
             return Err(NoteRecordError::StateTransitionError(
                 "Transaction ID does not match the expected value".to_string(),
@@ -75,10 +67,8 @@ impl NoteStateHandler for ProcessingAuthenticatedNoteState {
         }
 
         Ok(Some(
-            ConsumedAuthenticatedLocalNoteState {
+            ConsumedUnauthenticatedLocalNoteState {
                 metadata: self.metadata,
-                inclusion_proof: self.inclusion_proof.clone(),
-                block_note_root: self.block_note_root,
                 nullifier_block_height: block_height,
                 submission_data: self.submission_data,
             }
@@ -91,7 +81,7 @@ impl NoteStateHandler for ProcessingAuthenticatedNoteState {
     }
 
     fn inclusion_proof(&self) -> Option<&NoteInclusionProof> {
-        Some(&self.inclusion_proof)
+        None
     }
 
     fn consumer_transaction_id(&self) -> Option<&TransactionId> {
@@ -99,34 +89,31 @@ impl NoteStateHandler for ProcessingAuthenticatedNoteState {
     }
 }
 
-impl miden_tx::utils::Serializable for ProcessingAuthenticatedNoteState {
+impl miden_tx::utils::Serializable for ProcessingUnauthenticatedNoteState {
     fn write_into<W: miden_tx::utils::ByteWriter>(&self, target: &mut W) {
         self.metadata.write_into(target);
-        self.inclusion_proof.write_into(target);
-        self.block_note_root.write_into(target);
+        self.after_block_num.write_into(target);
         self.submission_data.write_into(target);
     }
 }
 
-impl miden_tx::utils::Deserializable for ProcessingAuthenticatedNoteState {
+impl miden_tx::utils::Deserializable for ProcessingUnauthenticatedNoteState {
     fn read_from<R: miden_tx::utils::ByteReader>(
         source: &mut R,
     ) -> Result<Self, miden_tx::utils::DeserializationError> {
         let metadata = NoteMetadata::read_from(source)?;
-        let inclusion_proof = NoteInclusionProof::read_from(source)?;
-        let block_note_root = Digest::read_from(source)?;
+        let after_block_num = u32::read_from(source)?;
         let submission_data = NoteSubmissionData::read_from(source)?;
-        Ok(ProcessingAuthenticatedNoteState {
+        Ok(ProcessingUnauthenticatedNoteState {
             metadata,
-            inclusion_proof,
-            block_note_root,
+            after_block_num,
             submission_data,
         })
     }
 }
 
-impl From<ProcessingAuthenticatedNoteState> for NoteState {
-    fn from(state: ProcessingAuthenticatedNoteState) -> Self {
-        NoteState::ProcessingAuthenticated(state)
+impl From<ProcessingUnauthenticatedNoteState> for InputNoteState {
+    fn from(state: ProcessingUnauthenticatedNoteState) -> Self {
+        InputNoteState::ProcessingUnauthenticated(state)
     }
 }