+++ crates/rust-client/src/store/note_record/input_note_record/states/unverified.rs
@@ -1,67 +1,77 @@
 use alloc::string::ToString;
 
 use miden_objects::{
-    notes::{NoteId, NoteInclusionProof, NoteMetadata},
+    notes::{compute_note_hash, NoteId, NoteInclusionProof, NoteMetadata},
     transaction::TransactionId,
-    BlockHeader, Digest,
+    BlockHeader,
 };
 
 use super::{
-    ConsumedExternalNoteState, NoteState, NoteStateHandler, NoteSubmissionData,
-    ProcessingAuthenticatedNoteState,
+    CommittedNoteState, ConsumedExternalNoteState, InputNoteState, InvalidNoteState,
+    NoteStateHandler, NoteSubmissionData, ProcessingUnauthenticatedNoteState,
 };
 use crate::store::NoteRecordError;
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct CommittedNoteState {
+pub struct UnverifiedNoteState {
     /// Metadata associated with the note, including sender, note type, tag and other additional
     /// information.
     pub metadata: NoteMetadata,
-    /// Inclusion proof for the note inside the chain block.
+    /// Inclusion proof for the note inside the chain block. This proof is not yet verified.
     pub inclusion_proof: NoteInclusionProof,
-    /// Root of the note tree inside the block that verifies the note inclusion proof.
-    pub block_note_root: Digest,
 }
 
-impl NoteStateHandler for CommittedNoteState {
+impl NoteStateHandler for UnverifiedNoteState {
     fn inclusion_proof_received(
         &self,
         inclusion_proof: NoteInclusionProof,
         metadata: NoteMetadata,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
-        if self.inclusion_proof != inclusion_proof || self.metadata != metadata {
-            return Err(NoteRecordError::StateTransitionError(
-                "Inclusion proof or metadata do not match the expected values".to_string(),
-            ));
-        }
-        Ok(None)
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
+        Ok(Some(UnverifiedNoteState { metadata, inclusion_proof }.into()))
     }
 
     fn consumed_externally(
         &self,
         nullifier_block_height: u32,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Ok(Some(ConsumedExternalNoteState { nullifier_block_height }.into()))
     }
 
     fn block_header_received(
         &self,
-        _note_id: NoteId,
+        note_id: NoteId,
         block_header: BlockHeader,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
-        if block_header.note_root() != self.block_note_root {
-            return Err(NoteRecordError::StateTransitionError(
-                "Block header does not match the expected note root".to_string(),
-            ));
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
+        if self.inclusion_proof.note_path().verify(
+            self.inclusion_proof.location().node_index_in_block().into(),
+            compute_note_hash(note_id, &self.metadata),
+            &block_header.note_root(),
+        ) {
+            Ok(Some(
+                CommittedNoteState {
+                    inclusion_proof: self.inclusion_proof.clone(),
+                    metadata: self.metadata,
+                    block_note_root: block_header.note_root(),
+                }
+                .into(),
+            ))
+        } else {
+            Ok(Some(
+                InvalidNoteState {
+                    metadata: self.metadata,
+                    invalid_inclusion_proof: self.inclusion_proof.clone(),
+                    block_note_root: block_header.note_root(),
+                }
+                .into(),
+            ))
         }
-        Ok(None)
     }
 
     fn consumed_locally(
         &self,
         consumer_account: miden_objects::accounts::AccountId,
         consumer_transaction: miden_objects::transaction::TransactionId,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
         let submission_data = NoteSubmissionData {
             submitted_at: None,
             consumer_account,
@@ -69,10 +79,9 @@ impl NoteStateHandler for CommittedNoteState {
         };
 
         Ok(Some(
-            ProcessingAuthenticatedNoteState {
+            ProcessingUnauthenticatedNoteState {
                 metadata: self.metadata,
-                inclusion_proof: self.inclusion_proof.clone(),
-                block_note_root: self.block_note_root,
+                after_block_num: self.inclusion_proof.location().block_num() - 1,
                 submission_data,
             }
             .into(),
@@ -83,7 +92,7 @@ impl NoteStateHandler for CommittedNoteState {
         &self,
         _transaction_id: TransactionId,
         _block_height: u32,
-    ) -> Result<Option<NoteState>, NoteRecordError> {
+    ) -> Result<Option<InputNoteState>, NoteRecordError> {
         Err(NoteRecordError::InvalidStateTransition(
             "Only processing notes can be committed in a local transaction".to_string(),
         ))
@@ -102,31 +111,25 @@ impl NoteStateHandler for CommittedNoteState {
     }
 }
 
-impl miden_tx::utils::Serializable for CommittedNoteState {
+impl miden_tx::utils::Serializable for UnverifiedNoteState {
     fn write_into<W: miden_tx::utils::ByteWriter>(&self, target: &mut W) {
         self.metadata.write_into(target);
         self.inclusion_proof.write_into(target);
-        self.block_note_root.write_into(target);
     }
 }
 
-impl miden_tx::utils::Deserializable for CommittedNoteState {
+impl miden_tx::utils::Deserializable for UnverifiedNoteState {
     fn read_from<R: miden_tx::utils::ByteReader>(
         source: &mut R,
     ) -> Result<Self, miden_tx::utils::DeserializationError> {
         let metadata = NoteMetadata::read_from(source)?;
         let inclusion_proof = NoteInclusionProof::read_from(source)?;
-        let block_note_root = Digest::read_from(source)?;
-        Ok(CommittedNoteState {
-            metadata,
-            inclusion_proof,
-            block_note_root,
-        })
+        Ok(UnverifiedNoteState { metadata, inclusion_proof })
     }
 }
 
-impl From<CommittedNoteState> for NoteState {
-    fn from(state: CommittedNoteState) -> Self {
-        NoteState::Committed(state)
+impl From<UnverifiedNoteState> for InputNoteState {
+    fn from(state: UnverifiedNoteState) -> Self {
+        InputNoteState::Unverified(state)
     }
 }