+++ crates/rust-client/src/store/sqlite_store/sync.rs
@@ -1,15 +1,18 @@
 use alloc::{collections::BTreeSet, string::ToString, vec::Vec};
 
-use miden_objects::notes::{NoteInclusionProof, NoteTag};
+use miden_objects::notes::{NoteTag, Nullifier};
 use miden_tx::utils::{Deserializable, Serializable};
-use rusqlite::{named_params, params, Transaction};
+use rusqlite::{params, Transaction};
 
 use super::SqliteStore;
 use crate::{
     store::{
-        note_record::{NOTE_STATUS_COMMITTED, NOTE_STATUS_CONSUMED},
-        sqlite_store::{accounts::update_account, notes::upsert_input_note_tx},
-        CommittedNoteState, InputNoteRecord, NoteFilter, StoreError,
+        input_note_states::CommittedNoteState,
+        sqlite_store::{
+            accounts::update_account,
+            notes::{upsert_input_note_tx, upsert_output_note_tx},
+        },
+        InputNoteRecord, NoteFilter, OutputNoteRecord, StoreError,
     },
     sync::{NoteTagRecord, NoteTagSource, StateSyncUpdate},
 };
@@ -90,7 +93,8 @@ impl SqliteStore {
         &self,
         state_sync_update: StateSyncUpdate,
     ) -> Result<(), StoreError> {
-        let mut relevant_input_notes = self.get_relevant_sync_input_notes(&state_sync_update)?;
+        let (mut relevant_input_notes, mut relevant_output_notes) =
+            self.get_relevant_sync_notes(&state_sync_update)?;
 
         let StateSyncUpdate {
             block_header,
@@ -112,28 +116,13 @@ impl SqliteStore {
 
         // Update tracked output notes
         for (note_id, inclusion_proof) in committed_notes.updated_output_notes().iter() {
-            let block_num = inclusion_proof.location().block_num();
-            let note_index = inclusion_proof.location().node_index_in_block();
-
-            let inclusion_proof = NoteInclusionProof::new(
-                block_num,
-                note_index,
-                inclusion_proof.note_path().clone(),
-            )?
-            .to_bytes();
-
-            // Update output notes
-            const COMMITTED_OUTPUT_NOTES_QUERY: &str =
-                "UPDATE output_notes SET status = :status , inclusion_proof = :inclusion_proof WHERE note_id = :note_id";
-
-            tx.execute(
-                COMMITTED_OUTPUT_NOTES_QUERY,
-                named_params! {
-                    ":inclusion_proof": inclusion_proof,
-                    ":note_id": note_id.inner().to_hex(),
-                    ":status": NOTE_STATUS_COMMITTED.to_string(),
-                },
-            )?;
+            if let Some(output_note_record) =
+                relevant_output_notes.iter_mut().find(|n| n.id() == *note_id)
+            {
+                if output_note_record.inclusion_proof_received(inclusion_proof.clone())? {
+                    upsert_output_note_tx(&tx, output_note_record)?;
+                }
+            }
         }
 
         // Update tracked input notes
@@ -223,12 +212,14 @@ impl SqliteStore {
                 }
             }
 
-            const SPENT_OUTPUT_NOTE_QUERY: &str =
-                "UPDATE output_notes SET status = ?, nullifier_height = ? WHERE nullifier = ?";
-            tx.execute(
-                SPENT_OUTPUT_NOTE_QUERY,
-                params![NOTE_STATUS_CONSUMED.to_string(), block_num, nullifier.to_hex()],
-            )?;
+            if let Some(output_note_record) = relevant_output_notes
+                .iter_mut()
+                .find(|n| n.nullifier().is_some_and(|n| n == nullifier))
+            {
+                if output_note_record.nullifier_received(nullifier, block_num)? {
+                    upsert_output_note_tx(&tx, output_note_record)?;
+                }
+            }
         }
 
         Self::insert_block_header_tx(&tx, block_header, new_mmr_peaks, block_has_relevant_notes)?;
@@ -253,12 +244,12 @@ impl SqliteStore {
         Ok(())
     }
 
-    /// Get the input notes from the store that are relevant to the state sync update. Secifically,
+    /// Get the notes from the store that are relevant to the state sync update. Secifically,
     /// notes that were updated and nullified during the sync.
-    fn get_relevant_sync_input_notes(
+    fn get_relevant_sync_notes(
         &self,
         state_sync_update: &StateSyncUpdate,
-    ) -> Result<Vec<InputNoteRecord>, StoreError> {
+    ) -> Result<(Vec<InputNoteRecord>, Vec<OutputNoteRecord>), StoreError> {
         let StateSyncUpdate { nullifiers, synced_new_notes, .. } = state_sync_update;
 
         let updated_input_note_ids = synced_new_notes
@@ -266,18 +257,33 @@ impl SqliteStore {
             .iter()
             .map(|input_note| input_note.id())
             .collect::<Vec<_>>();
-        let updated_input_notes = self.get_input_notes(NoteFilter::List(updated_input_note_ids))?;
 
-        let nullifiers = nullifiers
+        let updated_output_note_ids = synced_new_notes
+            .updated_output_notes()
+            .iter()
+            .map(|output_note| output_note.0)
+            .collect::<Vec<_>>();
+
+        let updated_input_notes = self.get_input_notes(NoteFilter::List(updated_input_note_ids))?;
+        let updated_output_notes =
+            self.get_output_notes(NoteFilter::List(updated_output_note_ids))?;
+
+        let nullifiers: Vec<Nullifier> = nullifiers
             .iter()
             .map(|nullifier_update| nullifier_update.nullifier)
             .collect::<Vec<_>>();
-        let nullified_notes = self.get_input_notes(NoteFilter::Nullifiers(nullifiers))?;
 
-        let mut relevant_notes = updated_input_notes;
-        relevant_notes.extend(nullified_notes);
+        let nullified_input_notes =
+            self.get_input_notes(NoteFilter::Nullifiers(nullifiers.clone()))?;
+        let nullified_output_notes = self.get_output_notes(NoteFilter::Nullifiers(nullifiers))?;
 
-        Ok(relevant_notes)
+        let mut relevant_input_notes = updated_input_notes;
+        relevant_input_notes.extend(nullified_input_notes);
+
+        let mut relevant_output_notes = updated_output_notes;
+        relevant_output_notes.extend(nullified_output_notes);
+
+        Ok((relevant_input_notes, relevant_output_notes))
     }
 }